- formatted the file and few minor changes to the naming convention of files
