+++ generator-jhipster-gomicro/generators/server/templates/src/controllers/base.go
@@ -1,7 +1,7 @@
 package controllers
 
 import (
-	"github.com/gorilla/mux"
+ "github.com/gorilla/mux"
 )
 
 type Controller interface {
