+++ generator-jhipster-gomicro/generators/server/templates/src/controllers/management.go
