+++ generator-jhipster-gomicro/generators/server/templates/src/pb/db.pb.micro.go
@@ -4,9 +4,9 @@
 package db
 
 import (
-	fmt "fmt"
-	proto "github.com/golang/protobuf/proto"
-	math "math"
+ fmt "fmt"
+ proto "github.com/golang/protobuf/proto"
+ math "math"
 )
 
 // Reference imports to suppress errors if they are not otherwise used.
