+++ generator-jhipster-gomicro/generators/server/templates/src/serviceregistry/discoverymanager.go
