+++ generator-jhipster-gomicro/generators/server/templates/src/serviceregistry/helper/httphelper.go
@@ -1,13 +1,13 @@
 package helper
 
 import (
-	"bytes"
-	"encoding/json"
-	"errors"
-	"net/http"
-	"time"
+ "bytes"
+ "encoding/json"
+ "errors"
+ "net/http"
+ "time"
 
-	"github.com/micro/micro/v3/service/logger"
+ "github.com/micro/micro/v3/service/logger"
 )
 
 type OperationBody interface {
