+++ generator-jhipster-gomicro/generators/server/templates/src/serviceregistry/helper/utility.go
@@ -1,8 +1,8 @@
 package helper
 
 import (
-	"errors"
-	"net"
+ "errors"
+ "net"
 )
 
 func ExternalIP() (string, error) {
