+++ generator-jhipster-gomicro/generators/server/templates/src/serviceregistry/registrationmanager.go
